Create a writer class for query results
Implementations of QueryResultWriter are intended to be used when
producing some type of formatted output for a query result.

The current implementions are for Verbose and Concise. The verbose
one prints out everything in the query result, formatted nicely with
titles for each field. The concise one, on the other hand, doesn't
have any titles and just prints out the pertitent information in a
single line.

My idea is that verbose might be better for debugging, while concise
is better for building something that just needs to know what was
affected, how much, and how long, such as an email.