Abstract time logic and create a query result set
Pulled the timer code out of QueryResult and moved it up to its class,
TimedOperation.

Created a new class, QueryResultSet, which is just a group of query
results. A QueryResultSet also subclasses off TimedOperation, making
it easier to track the total duration of batch operations.

QueryResultSet also has a total_count() method, which will iterate
through all of its query results for a total rowcount.