Reorganize, refactor, rewrite and convert to python3
This is a pretty substantial commit.

First, all the modules were converted to python3. As part of that, the
files were reorganized for python3 style module and package imports
(no relative imports unless used with from). I admittedly had a pretty
low understanding of python's import statement before this restructure,
but with python3 forcing me to redo them all, I was also forced to grok
python import.

The Query module was renamed to models. There is also a new duration class
that is used to calculate and format durations in TimedOperations. Other
than that, the rest of the models are basically the same.

The database connections were enhanced a bit. For starters, everything has
been moved around and renamed to better fit python naming conventions.
When I initially wrote the code, I was unaware of the lower level
python dbapi that many packages seem to use. With this in mind, I
abstracted all of the PostgresDBConnection code into a higher level
DBConnection (the original DBConnection was renamed to
AbstractDBConnection) and designed its constructor to receive a
"driver" argument, which is intended to be a package that implements the
dbapi standards (like sqlite3 and psycopg2). Having all of this core
logic abstracted to a higher level means that the implementation
modules simply subclass and supply a driver to super. This is how the
postgres module works and I also setup a sqlite module. My reasoning
for the sqlite module is actually so that I can later leverage the
sqlite3 memory-only connection to unit test DBConnection.